refactor(useId): use plugin for injecting id generator function
