+++ packages/bootstrap-vue-next/src/utils/exports/index.ts
@@ -1,1 +1,1 @@
-export {BvEvent, BvTriggerableEvent, BvCarouselEvent, provideGetId} from '..'
+export {BvEvent, BvTriggerableEvent, BvCarouselEvent} from '..'
