+++ packages/bootstrap-vue-next/src/utils/index.ts
@@ -7,7 +7,6 @@ export {default as formatItem} from './formatItem'
 export {default as getBreakpointProps} from './getBreakpointProps'
 export {default as getClasses} from './getClasses'
 export {default as getElement} from './getElement'
-export {default as getId, provideGetId} from './getId'
 export {default as getSlotElements} from './getSlotElements'
 export {default as getTableFieldHeadLabel} from './getTableFieldHeadLabel'
 export {default as isLink} from './isLink'
